Add DummyState class for use in offline mode.
Currently each state class separately implements support for keeping values in memory in offline mode. This change centralizes that behavior to a single class which will then be used instead of the real classes in the object's state register. The individual states' offline behavior support will be removed in the upcoming refactor of states.